requirements.txt
pdfplumber
Pillow
flask
pytesseract
